Refactor: Reconnect updated module. project. #107
